API: allow scalar setting/getting via float indexer on integer indexes
closes #12333

Author: Jeff Reback <jeff@reback.net>

Closes #12370 from jreback/float and squashes the following commits:

3a148a7 [Jeff Reback] API: allow scalar setting/getting via float indexer on integer indexes
