+++ pandas/tests/frame/test_indexing.py
@@ -216,7 +216,7 @@ class TestDataFrameIndexing(tm.TestCase, TestData):
         # we are producing a warning that since the passed boolean
         # key is not the same as the given index, we will reindex
         # not sure this is really necessary
-        with tm.assert_produces_warning(UserWarning):
+        with tm.assert_produces_warning(UserWarning, check_stacklevel=False):
             indexer_obj = indexer_obj.reindex(self.tsframe.index[::-1])
             subframe_obj = self.tsframe[indexer_obj]
             assert_frame_equal(subframe_obj, subframe)
